Code review: move test to the proper location
